+++ dao/workflow-node-instance.js
@@ -12,9 +12,25 @@
   "use strict";
 
 const {resolve} = require("path");
+const {throwParametersError} = require(resolve(__dirname, "..", "errors"));
 const commonsLib = require(resolve(__dirname, "public"))();
 
+const simpleInfo = {
+};
+
+const queryById = function* (id, hooks = {}) {
+  if (this.valid(id)) {
+    id = this.newObjectId(id);
+    if (hooks) {
+      Object.assign(hooks, simpleInfo);
+    }
+    return this.findOne({_id: id}, hooks);
+  }
+  throwParametersError();
+};
+
 const classMethod = {
+  queryById,
 };
 
 Object.assign(commonsLib, classMethod);