+++ routers/index.js
@@ -44,6 +44,7 @@ module.exports = _ => {
    */
   _.get("/workflow/list", authentication, workflow.list); // 工作流信息列表
   _.get("/workflow/:id/simple", authentication, workflow.simpleInfo); // 获取工作流实例简单信息
+  _.get("/workflow/:nodeId/files", authentication, workflow.nodeFiles); // 获取工作流某实例下的文件列表
 
   /*
    *  文件系统模块